+++ CHANGES.rst
@@ -5,6 +5,7 @@ The ASDF Standard is at v1.6.0
 
 - Overhaul of the ASDF documentation to make it more consistent and readable. [#1142]
 - Update deprecated instances of `abstractproperty` to `abstractmethod` [#1148]
+- Move build configuration into `pyproject.toml` [#1149]
 
 2.12.0 (2022-06-06)
 -------------------